StyTR-2-mindspore/requirements.txt
easydict==1.9
mindspore
numpy
Pillow